Removed the modbus library
It was decided that keeping the modbus code in a reusable library was
making the code too complex.  There is little need for a separate library
since there are many modbus libraries out there to choose from.  Allowing
the modbus code to be more closely coupled to the opendax code should
make the code easier to maintain.  Message logging will be more consistent.
Configuration code is easier and we can eliminate the callback functions
that we are using to interact with the tagbase.
