+++ src/modules/modbus/CMakeLists.txt
@@ -15,10 +15,8 @@
 #  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
 
 include_directories(.)
-add_subdirectory(lib)
-add_executable(modbus_module modmain.c modopt.c database.c)
+add_executable(modbus_module modmain.c modopt.c database.c mbcmds.c mbports.c mbserver.c mbslave.c mbutil.c modbus.c)
 set_target_properties(modbus_module PROPERTIES OUTPUT_NAME daxmodbus)
-target_link_libraries(modbus_module modbus_library)
 target_link_libraries(modbus_module dax)
 target_link_libraries(modbus_module pthread)
